+++ src/main/kotlin/controllers/worldRenderer/InputHandler.kt
@@ -37,7 +37,7 @@ class InputHandler internal constructor(
 
     fun tick(dt: Double) {
         if (dt > 1000) { // big lag spike, don't send the user flying
-            frameRateModel.needAnotherFrame = true
+            frameRateModel.notifyNeedFrames()
             return
         }
         val xVec = (-camera.yawSin).toDouble() / 65535
@@ -81,7 +81,7 @@ class InputHandler internal constructor(
         }
         if (motionTicked) {
             camera.motionTicks++
-            frameRateModel.needAnotherFrame = true
+            frameRateModel.notifyNeedFrames()
         }
     }
 
@@ -89,7 +89,7 @@ class InputHandler internal constructor(
         if (keys[code] == STATE_RELEASED) return false
         if (keys[code] == STATE_HELD || !frameRateModel.powerSavingMode.get()) return true
         keys[code] = STATE_HELD
-        frameRateModel.needAnotherFrame = true
+        frameRateModel.notifyNeedFrames()
         return false
     }