+++ src/main/kotlin/ui/CancelledException.kt
@@ -1,0 +1,3 @@
+package ui
+
+class CancelledException : Exception()