+++ src/main/kotlin/utils/ProgressContainer.kt
@@ -1,0 +1,9 @@
+package utils
+
+interface ProgressContainer {
+    var progress: Int
+    var progressMax: Int
+    var status: String
+    var isCancelled: Boolean
+    fun complete()
+}